Deploy heroku now

+++ src/main/java/com/kadias/demospringboot/services/UserServices.java
@@ -2,6 +2,11 @@ package com.kadias.demospringboot.services;
 
 import com.kadias.demospringboot.entities.User;
 import com.kadias.demospringboot.repositories.UserRepository;
+import com.kadias.demospringboot.services.exceptions.DatabaseExceptions;
+import com.kadias.demospringboot.services.exceptions.ResourceNotFoundException;
+import jakarta.persistence.EntityNotFoundException;
+import org.springframework.dao.DataIntegrityViolationException;
+import org.springframework.dao.EmptyResultDataAccessException;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
@@ -21,17 +26,41 @@ public class UserServices {
 
     public User findByID(Long id) {
         Optional<User> obj = userRepository.findById(id);
-        return obj.get();
+        return obj.orElseThrow(() -> new ResourceNotFoundException(id));
     }
 
     public User insert(User obj) {
         return userRepository.save(obj);
     }
 
-    public void delete (long id) {
-        userRepository.deleteById(id);
+    public void delete(long id) {
+        if (!userRepository.existsById(id)) {
+            throw new ResourceNotFoundException(id);
+        }
 
+        try {
+            userRepository.deleteById(id);
+        } catch (DataIntegrityViolationException e) {
+
+            throw new DatabaseExceptions(e.getMessage());
+        }
     }
 
+    public User update(long id, User obj) {
+        try {
+            User entity = userRepository.getReferenceById(id);
+
+            updateData(entity, obj);
+            return userRepository.save(entity);
+        } catch (EntityNotFoundException e) {
+            throw new ResourceNotFoundException(id);
+        }
+    }
+
+    private void updateData(User entity, User obj) {
+        entity.setName(obj.getName());
+        entity.setEmail(obj.getEmail());
+        entity.setPhone(obj.getPhone());
+    }
 
 }
